+++ src/ng-select/lib/config.service.ts
@@ -10,4 +10,6 @@ export class NgSelectConfig {
     clearAllText = 'Clear all';
     disableVirtualScroll = true;
     openOnEnter = true;
+    appendTo: string;
+    bindValue: string;
 }