+++ src/ng-select/lib/ng-select.component.spec.ts
@@ -9,6 +9,7 @@ import { MockConsole, MockNgZone } from '../testing/mocks';
 import { NgSelectComponent } from '@ng-select/ng-select';
 import { NgSelectModule } from './ng-select.module';
 import { Subject } from 'rxjs';
+import { NgSelectConfig } from './config.service';
 
 describe('NgSelectComponent', () => {
 
@@ -154,6 +155,86 @@ describe('NgSelectComponent', () => {
                         [clearable]="true"
                         [(ngModel)]="selectedCity">
                 </ng-select>`);
+
+            fixture.componentInstance.cities = [];
+            fixture.componentInstance.selectedCity = { id: 7, name: 'Pailgis' };
+            tickAndDetectChanges(fixture);
+
+            fixture.componentInstance.cities = [{ id: 7, name: 'Pailgis' }];
+            tickAndDetectChanges(fixture);
+
+            select = fixture.componentInstance.select;
+            expect(select.selectedItems[0]).toBe(select.itemsList.items[0]);
+            expect(select.selectedItems).toEqual([jasmine.objectContaining({
+                value: { id: 7, name: 'Pailgis' }
+            })]);
+        }));
+
+        it('should set items correctly after ngModel set first when bindValue is used from NgSelectConfig', fakeAsync(() => {
+            const config = new NgSelectConfig();
+            config.bindValue = 'id';
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="cities"
+                        bindLabel="name"
+                        [clearable]="true"
+                        [(ngModel)]="selectedCityId">
+                </ng-select>`,
+                config
+                );
+
+            fixture.componentInstance.cities = [];
+            fixture.componentInstance.selectedCityId = 7;
+            tickAndDetectChanges(fixture);
+
+            fixture.componentInstance.cities = [{ id: 7, name: 'Pailgis' }];
+            tickAndDetectChanges(fixture);
+
+            select = fixture.componentInstance.select;
+            expect(select.selectedItems[0]).toBe(select.itemsList.items[0]);
+            expect(select.selectedItems).toEqual([jasmine.objectContaining({
+                value: { id: 7, name: 'Pailgis' }
+            })]);
+        }));
+
+        it('should not apply global bindValue from NgSelectConfig if bindValue prop explicitly provided in template', fakeAsync(() => {
+            const config = new NgSelectConfig();
+            config.bindValue = 'globalbindvalue';
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="cities"
+                        bindLabel="name"
+                        bindValue="id"
+                        [clearable]="true"
+                        [(ngModel)]="selectedCityId">
+                </ng-select>`,
+                config
+                );
+
+            fixture.componentInstance.cities = [];
+            fixture.componentInstance.selectedCityId = 7;
+            tickAndDetectChanges(fixture);
+
+            fixture.componentInstance.cities = [{ id: 7, name: 'Pailgis' }];
+            tickAndDetectChanges(fixture);
+
+            select = fixture.componentInstance.select;
+            expect(select.selectedItems[0]).toBe(select.itemsList.items[0]);
+            expect(select.selectedItems).toEqual([jasmine.objectContaining({
+                value: { id: 7, name: 'Pailgis' }
+            })]);
+        }));
+
+        it('should bind whole object as value when bindValue prop is specified with empty string in template', fakeAsync(() => {
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="cities"
+                        bindLabel="name"
+                        bindValue=""
+                        [clearable]="true"
+                        [(ngModel)]="selectedCity">
+                </ng-select>`
+                );
 
             fixture.componentInstance.cities = [];
             fixture.componentInstance.selectedCity = { id: 7, name: 'Pailgis' };
@@ -431,7 +512,7 @@ describe('NgSelectComponent', () => {
 
 
             const disabled = { ...fixture.componentInstance.cities[1], disabled: true };
-            fixture.componentInstance.selectedCities = <any> [fixture.componentInstance.cities[0], disabled];
+            fixture.componentInstance.selectedCities = <any>[fixture.componentInstance.cities[0], disabled];
             tickAndDetectChanges(fixture);
 
             fixture.componentInstance.cities[1].disabled = true;
@@ -1445,7 +1526,7 @@ describe('NgSelectComponent', () => {
                 const remove = spyOn(select.removeEvent, 'emit');
                 fixture.componentInstance.multiple = true;
                 const disabled = { ...fixture.componentInstance.cities[1], disabled: true };
-                fixture.componentInstance.selectedCity = <any> [fixture.componentInstance.cities[0], disabled];
+                fixture.componentInstance.selectedCity = <any>[fixture.componentInstance.cities[0], disabled];
                 tickAndDetectChanges(fixture);
                 fixture.componentInstance.cities[1].disabled = true;
                 fixture.componentInstance.cities = [...fixture.componentInstance.cities];
@@ -2258,6 +2339,11 @@ describe('NgSelectComponent', () => {
                 tickAndDetectChanges(fixture);
                 expect(select.showAddTag).toBeFalsy();
             }));
+
+            it('should be false when there is search term with only empty space', () => {
+                select.searchTerm = '   ';
+                expect(select.showAddTag).toBeFalsy();
+            });
         });
     });
 
@@ -3070,7 +3156,7 @@ describe('NgSelectComponent', () => {
 
                 spyOn(fixture.componentInstance, 'onChange');
                 const disabled = { ...fixture.componentInstance.cities[1], disabled: true };
-                fixture.componentInstance.selectedCities = <any> [fixture.componentInstance.cities[0], disabled];
+                fixture.componentInstance.selectedCities = <any>[fixture.componentInstance.cities[0], disabled];
                 tickAndDetectChanges(fixture);
                 fixture.componentInstance.cities[1].disabled = true;
                 fixture.componentInstance.cities = [...fixture.componentInstance.cities];
@@ -3237,6 +3323,53 @@ describe('NgSelectComponent', () => {
                 expect(dropdown.style.left).toBe('0px');
             });
         }));
+
+        it('should apply global appendTo from NgSelectConfig', async(() => {
+            const config = new NgSelectConfig();
+            config.appendTo = 'body';
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `
+                <div class="container"></div>
+                <ng-select [items]="cities"
+                        [(ngModel)]="selectedCity">
+                </ng-select>`,
+                config);
+
+            fixture.componentInstance.select.open();
+            fixture.detectChanges();
+
+            fixture.whenStable().then(() => {
+                const dropdown = <HTMLElement>document.querySelector('.ng-dropdown-panel');
+                expect(dropdown.parentElement).toBe(document.body);
+                expect(dropdown.style.top).not.toBe('0px');
+                expect(dropdown.style.left).toBe('0px');
+            });
+        }));
+
+        it('should not apply global appendTo from NgSelectConfig if appendTo prop explicitly provided in template', async(() => {
+            const config = new NgSelectConfig();
+            config.appendTo = 'body';
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `
+                <div class="container"></div>
+                <ng-select [items]="cities"
+                        appendTo=".container"
+                        [(ngModel)]="selectedCity">
+                </ng-select>`,
+                config);
+
+            fixture.componentInstance.select.open();
+            fixture.detectChanges();
+
+            fixture.whenStable().then(() => {
+                const dropdown = <HTMLElement>document.querySelector('.container .ng-dropdown-panel');
+                expect(dropdown.style.top).not.toBe('0px');
+                expect(dropdown.style.left).toBe('0px');
+            });
+        }));
+
     });
 
     describe('Grouping', () => {
@@ -3506,10 +3639,61 @@ describe('NgSelectComponent', () => {
             }));
         });
     });
+
+    describe('User defined keyDown handler', () => {
+        let fixture: ComponentFixture<NgSelectTestCmp>;
+        let select: NgSelectComponent;
+
+        beforeEach(() => {
+            fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [keyDownFn]="keyDownFn"></ng-select>`);
+            select = fixture.componentInstance.select;
+        });
+
+        const expectSpyToBeCalledAfterKeyDown = (spy, expectedNumber) => {
+            const possibleKeyCodes = Object.keys(KeyCode);
+            possibleKeyCodes
+                .forEach(keyCode => {
+                    triggerKeyDownEvent(getNgSelectElement(fixture), KeyCode[keyCode])
+                });
+            expect(spy).toHaveBeenCalledTimes(expectedNumber)
+        };
+
+        it('should execute user function if any of defined keys was pressed', () => {
+            const spy = spyOn(fixture.componentInstance.select, 'keyDownFn');
+            expectSpyToBeCalledAfterKeyDown(spy, Object.keys(KeyCode).length)
+        });
+
+        it('should not call any of default keyDown handlers if user handler returns false', fakeAsync(() => {
+            fixture.componentInstance.keyDownFn = () => false;
+            tickAndDetectChanges(fixture);
+            const spy = spyOn(fixture.componentInstance.select, 'handleKeyCode');
+
+            expectSpyToBeCalledAfterKeyDown(spy, 0)
+        }));
+
+        it('should call default keyHandler if user handler returns truthy', fakeAsync(() => {
+            fixture.componentInstance.keyDownFn = () => true;
+            tickAndDetectChanges(fixture);
+
+            const spy = spyOn(fixture.componentInstance.select, 'handleKeyCode');
+            expectSpyToBeCalledAfterKeyDown(spy, Object.keys(KeyCode).length)
+        }));
+
+        it('should call default keyHandler if user handler returns falsy but not `false`', fakeAsync(() => {
+            fixture.componentInstance.keyDownFn = () => null;
+            tickAndDetectChanges(fixture);
+
+            const spy = spyOn(fixture.componentInstance.select, 'handleKeyCode');
+            expectSpyToBeCalledAfterKeyDown(spy, Object.keys(KeyCode).length)
+        }))
+    })
 });
 
 
-function createTestingModule<T>(cmp: Type<T>, template: string): ComponentFixture<T> {
+function createTestingModule<T>(cmp: Type<T>, template: string, customNgSelectConfig: NgSelectConfig | null = null): ComponentFixture<T> {
+
     TestBed.configureTestingModule({
         imports: [FormsModule, NgSelectModule],
         declarations: [cmp],
@@ -3523,8 +3707,13 @@ function createTestingModule<T>(cmp: Type<T>, template: string): ComponentFixtur
             set: {
                 template: template
             }
-        })
-        .compileComponents();
+        });
+
+        if (customNgSelectConfig) {
+            TestBed.overrideProvider(NgSelectConfig, { useValue: customNgSelectConfig });
+        }
+
+        TestBed.compileComponents();
 
     const fixture = TestBed.createComponent(cmp);
     fixture.detectChanges();
@@ -3582,6 +3771,7 @@ class NgSelectTestCmp {
         { id: 2, description: { name: 'USA', id: 'b' } },
         { id: 3, description: { name: 'Australia', id: 'c' } }
     ];
+    keyDownFn = () => {};
 
     tagFunc(term: string) {
         return { id: term, name: term, custom: true }