+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### Features
 
-* **keyDownFn** user defined keydown handler to prevent ng-select behaviour([8b2b817](https://github.com/ng-select/ng-select/commit/8b2b817)), closes [#1155](https://github.com/ng-select/ng-select/issues/1155)
+* **keyDownFn:** user defined keydown handler to prevent ng-select behaviour([8b2b817](https://github.com/ng-select/ng-select/commit/8b2b817)), closes [#1155](https://github.com/ng-select/ng-select/issues/1155)
 
 
 ### Bug Fixes